[Refactor] DELETE request changes
- Changed path to add parameter
- Changed where query to make it the user ID. Will need to work with the session tokens on this eventually

+++ src/api/lib/routes/user.ts
@@ -268,14 +268,14 @@ router.post('/user/changeemail', async (req, res) => {
 })
 
 
-router.post('/users/deleteuser', async (req, res) => {
+router.delete('/users/:userID', async (req, res) => {
     try {
         const connection = await connect();
         await connection
             .createQueryBuilder()
             .delete()
             .from(User)
-            .where("email= :email", { email: req.body.email })
+            .where("id = :id", { id: req.params.userID })
             .execute();
     } catch (err) {
         console.log(err);
